Add OpenTelemetry support for AWS and Datadog
Adds [AWS Distro for OpenTelemetry](https://aws.amazon.com/otel/) (ADOT) support to the Kubernetes
cluster, using the [Amazon EKS add-on for ADOT] add-on. This adds cert-manager with a default
configuration and the OpenTelemetry Operator via the add-on. The Kubernetes service account is enabled, via OIDC, to assume an IAM role in AWS to upload traces.

The collector defaults to sending traces to AWS X-Ray, however Datadog support can also be enabled
by setting these keys in Pulumi config:

* `datadogEnabled` to `true`
* `datadogApiKey`

The API key must be obtained from the Datadog console.